scope option for ros_ws refresh

+++ turtlebot3/diamond_tools/ros_ws.sh
@@ -2,17 +2,20 @@
 
 ROS_WS="/home/ros/tb3_ws"
 REPO_DIR="/home/ros/repos"
+SCOPE=$1
 
-rm -rf $ROS_WS
-# Make a new workspace:
-mkdir -p $ROS_WS/src
+if [[ "$SCOPE" == "full" || "$SCOPE" == "f" ]]; then
+  rm -rf $ROS_WS
+  # Make a new workspace:
+  mkdir -p $ROS_WS/src
 
-cd $REPO_DIR/turtlebot3/
-# git pull --quiet
-tb3_pkgs=("bringup" "description" "node" "teleop")
-for tb3_pkg in ${tb3_pkgs[@]}; do
-  cp -r turtlebot3_$tb3_pkg $ROS_WS/src/
-done
+  cd $REPO_DIR/turtlebot3/
+  # git pull --quiet
+  tb3_pkgs=("bringup" "description" "node" "teleop")
+  for tb3_pkg in ${tb3_pkgs[@]}; do
+    cp -r turtlebot3_$tb3_pkg $ROS_WS/src/
+  done
+fi
 
 cd $REPO_DIR/tuos_ros/
 git pull --quiet
@@ -22,5 +25,8 @@ for tuos_pkg in ${tuos_pkgs[@]}; do
   cp -r tuos_$tuos_pkg $ROS_WS/src/
 done
 
-cd $ROS_WS/ && colcon --log-level ERROR build --symlink-install
-# cd $ROS_WS/ && colcon --log-level ERROR build --packages-select tuos_tb3_tools --symlink-install
+if [[ "$SCOPE" == "full" || "$SCOPE" == "f" ]]; then
+  cd $ROS_WS/ && colcon --log-level ERROR build --symlink-install
+else
+  cd $ROS_WS/ && colcon --log-level ERROR build --packages-select tuos_tb3_tools --symlink-install
+fi
